requirements.txt
Package                        Version
------------------------------ ---------------
absl-py                        0.15.0
actionlib-msgs                 1.13.0.post3
adversarial-robustness-toolbox 1.10.0
aif360                         0.4.0
argon2-cffi                    21.3.0
argon2-cffi-bindings           21.2.0
art                            5.7
astor                          0.8.1
astunparse                     1.6.3
async-generator                1.10
attrs                          21.4.0
audioread                      2.1.9
backcall                       0.2.0
backports.weakref              1.0rc1
BlackBoxAuditing               0.1.54
bleach                         1.5.0
boto                           2.49.0
cached-property                1.5.2
cachetools                     4.2.4
catkin                         0.7.18
catkin-pkg                     0.5.2
certifi                        2021.5.30
cffi                           1.15.0
charset-normalizer             2.0.12
clang                          5.0
cloudpickle                    2.0.0
convex-adversarial             0.4.4
cv-bridge                      1.13.0.post0
cvxpy                          1.1.19
cycler                         0.11.0
dataclasses                    0.8
decorator                      4.4.2
defusedxml                     0.7.1
docutils                       0.18.1
eagerpy                        0.30.0
ecos                           2.0.10
entrypoints                    0.4
enum34                         1.1.10
fairlearn                      0.4.6
flatbuffers                    1.12
foolbox                        3.3.3
gast                           0.4.0
genmsg                         0.5.12
genpy                          0.6.14
geometry-msgs                  1.13.0.post2
gitdb                          4.0.9
GitPython                      3.1.20
gnupg                          2.3.1
google-auth                    1.35.0
google-auth-oauthlib           0.4.6
google-pasta                   0.2.0
gpuinfo                        1.0.0a7
grpcio                         1.44.0
h5py                           3.1.0
html5lib                       0.9999999
idna                           3.3
imageio                        2.15.0
importlib-metadata             4.8.3
importlib-resources            5.4.0
iniconfig                      1.1.1
ipykernel                      5.5.6
ipython                        7.16.3
ipython-genutils               0.2.0
ipywidgets                     7.7.0
jedi                           0.17.2
Jinja2                         3.0.3
joblib                         1.1.0
jsonschema                     4.0.0
jupyter                        1.0.0
jupyter-client                 7.1.2
jupyter-console                6.4.3
jupyter-core                   4.9.2
jupyterlab-pygments            0.1.2
jupyterlab-widgets             1.1.0
jupyterthemes                  0.20.0
keras                          2.6.0
Keras-Applications             1.0.8
Keras-Preprocessing            1.1.2
kiwisolver                     1.3.1
lesscpy                        0.15.0
librosa                        0.5.0
lime                           0.2.0.1
llvmlite                       0.36.0
Markdown                       3.3.6
MarkupSafe                     2.0.1
matplotlib                     3.3.4
memory-profiler                0.60.0
mistune                        0.8.4
mkl-fft                        1.3.0
mkl-random                     1.1.1
mkl-service                    2.3.0
nav-msgs                       1.13.0.post3
nbclient                       0.5.9
nbconvert                      6.0.7
nbformat                       5.1.3
nest-asyncio                   1.5.4
networkx                       2.5.1
notebook                       6.4.10
numba                          0.53.1
numpy                          1.19.5
oauthlib                       3.2.0
olefile                        0.46
opencv-python                  4.5.5.64
opt-einsum                     3.3.0
osqp                           0.6.2.post5
packaging                      21.3
pandas                         1.1.5
pandocfilters                  1.5.0
parso                          0.7.1
pexpect                        4.8.0
pickleshare                    0.7.5
Pillow                         8.4.0
pip                            21.2.2
pluggy                         1.0.0
ply                            3.11
prettytable                    2.5.0
progressbar                    2.5
prometheus-client              0.13.1
prompt-toolkit                 3.0.28
protobuf                       3.19.4
psutil                         5.9.0
ptyprocess                     0.7.0
py                             1.11.0
pyasn1                         0.4.8
pyasn1-modules                 0.2.8
pycparser                      2.21
pycryptodome                   3.15.0
pycryptodomex                  3.15.0
pyecharts                      1.9.1
Pygments                       2.11.2
pyparsing                      3.0.6
pyrsistent                     0.18.0
pytest                         7.0.1
python-dateutil                2.8.2
pytz                           2021.3
PyWavelets                     1.1.1
PyYAML                         6.0
pyzmq                          22.3.0
qdldl                          0.1.5.post0
qtconsole                      5.2.2
QtPy                           2.0.1
requests                       2.27.1
requests-oauthlib              1.3.1
resampy                        0.2.2
ros-numpy                      0.0.2
rosbag                         1.15.11
roscpp                         1.15.11
rosgraph                       1.15.11
rosgraph-msgs                  1.11.3.post2
roslib                         1.14.7.post0
roslz4                         1.14.3.post2
rospkg                         1.4.0
rospy                          1.15.11
rsa                            4.8
scikit-image                   0.17.2
scikit-learn                   0.24.2
scipy                          1.5.4
scs                            3.2.0
seaborn                        0.11.2
Send2Trash                     1.8.0
sensor-msgs                    1.13.0.post3
setproctitle                   1.2.3
setuptools                     59.6.0
shap                           0.40.0
simplejson                     3.17.6
six                            1.15.0
sklearn                        0.0
slicer                         0.0.7
smmap                          5.0.0
std-msgs                       0.5.13.post0
tempeh                         0.1.12
tensorboard                    2.6.0
tensorboard-data-server        0.6.1
tensorboard-plugin-wit         1.8.1
tensorflow-estimator           2.6.0
tensorflow-federated           0.4.0
tensorflow-gpu                 2.6.2
tensorflow-tensorboard         1.5.1
termcolor                      1.1.0
terminado                      0.13.0
testpath                       0.6.0
tf2-msgs                       0.7.2.post3
threadpoolctl                  3.0.0
tifffile                       2020.9.3
tomli                          1.2.3
torch                          1.7.1
torchaudio                     0.7.0a0+a853dff
torchsummary                   1.5.1
torchvision                    0.8.2
tornado                        6.1
tqdm                           4.63.0
traitlets                      4.3.3
typing-extensions              3.7.4.3
urllib3                        1.26.9
visualization-msgs             1.13.0.post4
wcwidth                        0.2.5
webencodings                   0.5.1
Werkzeug                       2.0.3
wheel                          0.37.0
widgetsnbextension             3.6.0
wrapt                          1.12.1
xgboost                        1.5.2
z3                             0.2.0
zipp                           3.6.0